LayoutTests/media/track/track-mutable-expected.txt

Tests MutableTextTrack functionality: addTrack(), addCue(), and removeCue().


*** Testing addTrack()
RUN(track = video.addTrack("captions", "Mutable Captions Tract", "en"))
EXPECTED (video.textTracks.length == '1') OK

RUN(video.addTrack("subtitles"))
EXPECTED (video.textTracks.length == '2') OK

*** Testing addCue()
EXPECTED (track.cues.length == '0') OK
RUN(track.addCue("junk"))
EXPECTED (track.cues.length == '0') OK

RUN(track.addCue(new TextTrackCue("1", 1.2, 3.4, "Cue #1")))
EXPECTED (track.cues.length == '1') OK

RUN(track.addCue(new TextTrackCue("2", 5.6, 7.8, "Cue #2", "A:middle T:50%", true)))
EXPECTED (track.cues.length == '2') OK
EXPECTED (track.cues[0].getCueAsSource() == 'Cue #1') OK
EXPECTED (track.cues[1].getCueAsSource() == 'Cue #2') OK
EXPECTED (track.cues[1].pauseOnExit == 'true') OK

*** Testing removeCue()
RUN(track.removeCue(track.cues[1]))
EXPECTED (track.cues.length == '1') OK
EXPECTED (track.cues[0].getCueAsSource() == 'Cue #1') OK

RUN(track.removeCue(new TextTrackCue()))
EXPECTED (track.cues.length == '1') OK

RUN(track.removeCue(track.cues[0]))
EXPECTED (track.cues.length == '0') OK
END OF TEST
